refactor: modularize fact-checking tools and streamline exports
- Re-exported all functions from the fact-checking module for better organization and accessibility.
- Removed redundant code and improved the structure of the fact-checking tools.
- Enhanced clarity and maintainability of the codebase by consolidating related functionalities.
